Add clusterState to support operations on entire Elasticsearch cluster
ElasticsearchRequiredAction defines what action must be taken on the cluster
if any. The cluster is viewed as a single entity, rather than individually
node by node. This is required for making cluster-wide operations like rolling
restart of nodes.

clusterState struct holds the state of Elasticsearch cluster.
clsuterState gets constructed on each reconciliation.
clusterState contains information about the desired state of the cluster
mapped to the actual state of the cluster.

clusterState.getRequiredAction() returns the action that needs to be
performed on the cluster based on the clusterState.
At the moment 2 actions are supported: ElasticsearchActionNewClusterNeeded and
ElasticsearchActionScaleDownNeeded.

Added help for various storage options.