Support Rolling Restart on Image Change
In order to support minor releases of OpenShift logging
users will need to run rolling restart of the cluster.
This restart can be triggered bu updating
`ElasticsearchNodeSpec.Image`.

A prerequisite is that the cluster health is in green state.

This rolling restart will upgrade the cluster in a node-by-node fashion,
always one node at a time.

The operators waits for shard reallocation after every node upgrade,
thus a full upgrade procedure can take significat time to finish.

Synced flush is used to lower the total time that it takes
for the whole cluster to upgrade and reballance shards.

If the cluster restart procedure fails, a human intervention
is needed to bring the cluster back to operation.