Introduce gosec for security checks
gosec does static code analysis and checks for common security issues in
golang codebases.

This PR introduces the tool, and sets it up as a test target in the Makefile.

This also makes the creation of the folder with certificates/keys more
restrictive. This was one of the warnings found by gosec.