Bug 1857826: Assume average log message is 1K
The IndexManagement is configured to rollover index if any of existing conditions is met. We have
conditions on index size, number of document and age. Currently the condition on number of documents
in the index was calculated assuming the size of individual document (a single log) is up to 1M.
This is oversized for most of the real cases and is causing the index to be rolled over when 120.000
documents are stored in the index (given 3/1 prim/replicas sharding). This can easily lead to 10 new
infra indices per hour on an "idle" OCP cluster.

We are going to update the condition to assume average size of individual log is a lot less (~1K).