[release-4.5] Bug 1860135: Assume average log message is 1K
The IndexManagement is configured to rollover index if any of existing conditions is met. We have
conditions on index size, number of document and age. Currently the condition on number of documents
in the index was calculated assuming the size of individual document (a single log) is up to 1M.
This is oversized for most of the real cases and is causing the index to be rolled over when 120.000
documents are stored in the index (given 3/1 prim/replicas sharding). This can easily lead to 10 new
infra indices per hour (about 300mb size each) on an "idle" OCP cluster.

We update the condition to assume average size of individual log is a lot less (~1K). But in general this is not about assumed size of individual log message but about increasing the number of documents stored in index before maxSize rollover condition is met.