only alert for ES being red if the csv succeeded
During the rollout of a new logging-operator version, ES can take some
time before it goes green again. This patch only causes the alert for ES
being red to fire if the elasticsearch-operator csv succeeded.

If the csv is actively rolling out, ES being red should be expected and
we have other monitoring for when a csv is abnormal.